# repo/backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Meridian Wellness API
  version: 1.0.0
servers:
  - url: https://localhost:8443/api/v1
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
  /auth/register:
    post:
      summary: Register local user
      responses: { '201': { description: Created } }
  /auth/login:
    post:
      summary: Login user
      responses: { '200': { description: Token issued } }
  /auth/me:
    get:
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: Current user } }
  /catalog:
    get:
      responses: { '200': { description: Published packages and pricing calendar } }
  /config/coverage:
    get:
      responses: { '200': { description: Allowed coverage regions list } }
  /catalog/routes:
    get:
      responses: { '200': { description: Published routes } }
  /catalog/hotels:
    get:
      responses: { '200': { description: Published hotels } }
  /catalog/attractions:
    get:
      responses: { '200': { description: Published attractions } }
  /profile/addresses:
    get:
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: Traveler addresses } }
    post:
      security: [{ bearerAuth: [] }]
      responses: { '201': { description: Address added } }
  /profile/contacts:
    get:
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: Traveler contacts } }
    post:
      security: [{ bearerAuth: [] }]
      responses: { '201': { description: Contact added } }
  /profile/addresses/{id}:
    delete:
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses: { '200': { description: Deleted } }
  /profile/contacts/{id}:
    delete:
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses: { '200': { description: Deleted } }
  /bookings/holds:
    get:
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: Holds list } }
    post:
      security: [{ bearerAuth: [] }]
      responses: { '201': { description: Hold created }, '409': { description: Conflict } }
  /bookings/holds/{id}:
    delete:
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses: { '200': { description: Hold cancelled } }
  /bookings/history:
    get:
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: Booking history } }
  /bookings/confirm:
    post:
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: Hold confirmed into booking }, '409': { description: Version or state conflict } }
  /bookings/{id}/status:
    post:
      summary: Update booking status and optional clinical notes
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  description: Allowed values include scheduled, confirmed, checked_in, in_progress, completed, cancelled.
                notes:
                  type: string
                  description: Clinical or session notes saved encrypted at rest.
              required: [status]
      responses:
        '200': { description: Booking status updated }
        '400': { description: Invalid status or payload }
        '403': { description: Permission or ownership denied }
        '404': { description: Booking not found }
  /scheduling/slots:
    get:
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: hostId
          required: true
          schema: { type: integer }
        - in: query
          name: roomId
          required: true
          schema: { type: integer }
        - in: query
          name: day
          required: true
          schema: { type: string, format: date }
        - in: query
          name: duration
          required: true
          schema: { type: integer, enum: [30,45,60] }
      responses: { '200': { description: Available dynamic slots } }
  /scheduling/hosts/{id}/agenda:
    get:
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses: { '200': { description: Host agenda } }
  /scheduling/rooms/{id}/agenda:
    get:
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses: { '200': { description: Room agenda } }
  /community/posts:
    get:
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: Community posts } }
    post:
      security: [{ bearerAuth: [] }]
      responses: { '201': { description: Post created } }
  /community/posts/{id}/comments:
    get:
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses: { '200': { description: Threaded comments } }
    post:
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses: { '201': { description: Comment created } }
  /community/favorites:
    post:
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: Favorite added } }
  /community/likes:
    post:
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: Like added } }
  /community/follows:
    post:
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: Follow added } }
  /community/blocks:
    post:
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: Block added } }
  /community/reports:
    post:
      security: [{ bearerAuth: [] }]
      responses: { '201': { description: Moderation report created } }
  /notifications:
    get:
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: Notification feed } }
  /notifications/{id}/read:
    post:
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses: { '200': { description: Marked read } }
  /admin/roles/assign:
    post:
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: Role assigned + audited } }
  /admin/users:
    get:
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: User list by role } }
  /admin/roles/audits:
    get:
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: Permission audit feed } }
  /admin/reports/{id}/resolve:
    post:
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses: { '200': { description: Moderation resolved } }
  /ops/analytics/kpis:
    get:
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: KPI metrics with filters } }
  /ops/analytics/export:
    get:
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: KPI CSV export path } }
  /ops/reports:
    get:
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: Scheduled report job list } }
  /ops/reports/schedule:
    post:
      security: [{ bearerAuth: [] }]
      responses: { '201': { description: Local report job scheduled } }
  /ops/email/queue:
    get:
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: Internal email template queue } }
    post:
      security: [{ bearerAuth: [] }]
      responses: { '201': { description: Queue item created } }
  /ops/email/export:
    post:
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: Queue exported as local CSV } }